+++ src/textual/widgets/_input.py
@@ -129,7 +129,7 @@ class Input(ScrollView):
             show=False,
         ),
         Binding(
-            "ctrl+backspace",
+            "ctrl+backspace,alt+backspace",
             "delete_right_word",
             "Delete right to start of word",
             show=False,
@@ -150,6 +150,7 @@ class Input(ScrollView):
     | shift+right | Move cursor right and select. |
     | ctrl+right | Move the cursor one word to the right. |
     | backspace | Delete the character to the left of the cursor. |
+    | ctrl+backspace,alt+backspace | Delete right to start of word. |
     | ctrl+shift+right | Move cursor right a word and select. |
     | ctrl+shift+a | Select all text in the input. |
     | home,ctrl+a | Go to the beginning of the input. |